Creating RestaurantReviews page which retrieves reviews for a specified restaurant ID; moving homepage outside of App into separate component; initializing Routes for the application; installing React Redux for state management
