+++ frontend/package.json
@@ -11,6 +11,7 @@
     "react": "^18.2.0",
     "react-cookie": "^4.1.1",
     "react-dom": "^18.2.0",
+    "react-redux": "^8.1.1",
     "react-router-dom": "^5.3.0",
     "react-scripts": "5.0.1",
     "reactstrap": "^8.10.0",
