+++ frontend/src/App.js
@@ -1,36 +1,18 @@
-import { Component } from 'react';
-import logo from './FrenchFryFoodCritic.png';
-import './App.css';
+import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
+import Homepage from './components/Homepage';
+import RestaurantReviews from './components/RestaurantReviews';
 
-
-
-class App extends Component {
-
-  state = {
-    message: ""
-  };
-
-  async componentDidMount() {
-    const response = await fetch('/welcome');
-    const body = await response.text();
-    this.setState({message: body});
-  }
-
-  render() {
-    const {message} = this.state;
-    return (
-      <div className="App">
-        <header className="App-header">
-          <img src={logo} className="Food-critic" alt="logo" />
-            <div className="App-intro">
-              <h2>Welcome</h2>
-              <p>{message}</p>
-            </div>
-        </header>
-      </div>
-    );
-
-  }
+function App() {
+  return (
+    <div>
+      <Router>
+        <Switch>
+          <Route exact path="/" component={Homepage} />
+          <Route path="/reviews/:id" component={RestaurantReviews} />
+        </Switch>
+      </Router>
+    </div>
+  );
 }
 
 export default App;
